% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown.R
\docType{package}
\name{posterdown}
\alias{posterdown-package}
\alias{posterdown}
\title{posterdown package}
\description{
This package creates posters for conferences
}
\section{poster_html}{

Creates an R Markdown template for a PDF poster document via HTML
}

\section{poster_betterland}{

Creates an R Markdown template for a PDF poster document via HTML
}

\section{poster_betterport}{

Creates an R Markdown template for a PDF poster document via HTML
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brentthorne/posterdown}
  \item Report bugs at \url{https://github.com/brentthorne/posterdown/issues}
}

}
\author{
Brent Thorne
}
