% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown_generic.R
\name{posterdown_generic}
\alias{posterdown_generic}
\title{Posterdown PDF format}
\usage{
posterdown_generic(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Multidisciplinary Digital Publishing
Institute (MDPI) journals. Adapted from
\href{http://www.mdpi.com/authors/latex}{http://www.mdpi.com/authors/latex}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "posterdown_generic", package = "posterdown")
}

}
