% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown_html.R
\name{posterdown_html_gridlayout}
\alias{posterdown_html_gridlayout}
\title{Posterdown HTML format via grid layout (using pagedown::poster_relaxed)}
\usage{
posterdown_html_gridlayout(...,
  template = find_resource("posterdown_html_gridlayout",
  "template.html"), css = NULL)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::html_document}}

\item{template}{See \code{\link{html_paged}()}.}

\item{css}{See \code{\link{html_paged}()}.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Posterdown HTML format via grid layout (using pagedown::poster_relaxed)
}
