% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown_latex.R
\name{posterdown_latex}
\alias{posterdown_latex}
\alias{posterdown_pdf}
\title{Posterdown LaTeX format (using Memoir LaTeX document class)}
\usage{
posterdown_latex(..., keep_tex = TRUE)

posterdown_pdf(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Legacy function for the OG's of posterdown.
This will remain untouched from here on out (Feb 6 2019).
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "posterdown_pdf", package = "posterdown")
}
}
