% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown_pdf.R
\name{posterdown_pdf}
\alias{posterdown_pdf}
\title{Posterdown PDF format (using 'memoir' document class)}
\usage{
posterdown_pdf(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Posterdown PDF format (using 'memoir' document class)
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "posterdown_generic", package = "posterdown")
}

}
