% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown.R
\docType{package}
\name{posterdown}
\alias{posterdown}
\alias{posterdown-package}
\title{posterdown package}
\description{
This package creates posters for conferences
}
\section{poster_pdf}{

Creates an R Markdown template for a PDF poster document via LaTeX
}

\section{poster_html}{

Creates an R Markdown template for a PDF poster document via HTML
}

\author{
Brent Thorne
}
