% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown_html.R
\name{posterdown_html}
\alias{posterdown_html}
\title{Posterdown HTML format (using pagedown::poster_relaxed)}
\usage{
posterdown_html(
  ...,
  template = find_resource("posterdown_html", "template.html"),
  css = NULL
)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::html_document}}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Posterdown HTML format (using pagedown::poster_relaxed)
}
\examples{
\donttest{
file <- file.path(tempdir(),"foo.rmd")
rmarkdown::draft(file, template="posterdown_html", package="posterdown")
}

}
